Made base class more portable for PRO
This code was written in a different branch, but really belongs in the
same chain of code as the new renderer, so I'm combining it here.
Although technically ready to be deployed, this code should only be
used once PRO is available so that everything can be thoroughly tested
to work together.

- Try/Catch gallery instantiation to handle more complex errors
- Renamed variables to make code more readable
- We now create a cache token for every gallery, there is no sense in
storing normal cached data as a different convention (using the gallery
id as the cache ID) than data generated from an overloaded shortcode.
This means we use the cache code once instead of comparing the
shortcode settings to the settings stored in the database.
- Now uses renamed Pro_About class for page template. This means this
branch needs to be used in correlation with the pro/page branch.
- Minor readability changes to basic core