Changed data that is sent to several methods
Now we're only using what is necessary in each method instead of
sending along extra junk.

- Cache expiration can be NULL, meaning, if cache isn't set, get a
result set but don't cache it.
- Videos Endpoint is now set in shortcode
- Gallery ID from database is no longer sent to template
- Vimeo library now only adds params if they exist.
- Vimeo library client id was getting added to query string when I
didn't want the pagination query string to be modified. If Basic
requires this client id in the query string, simply check if it is set
and if so, append to query string manually.