Introduced the addons class
This class handles all of the registered Vimeography addons that hook in to the Vimeography plugin load with `do_action('vimeography/load-addon-plugin')`

This action uses the commented headers of the calling file to determine more information about the plugin.

This file also allows Vimeography to set the active Vimeography theme, so we know which one to use while rendering the galleries.