Add a method to clear out corrupt activation keys
This will remove a site option if a false positive is found in the database and add an admin notice asking the user to re-enter their activation keys. This could have been possible due to a bug in the previously used update library.

It also will move the activation keys into a site_option if it is a multi-site install.