Refactored the Gallery Editor class
The PHP was pretty crufty from 2 years of code, so I implemented the new vimeography_action system, which calls specific functions based on which key is submitted in a HTTP Request's GET or POST value. This removes the need to detect request parameters in a constructor, which makes me feel a lot better.