Add support for Vimeography Pro < 0.7
Vimeography 1.2 introduced a refactor that made Vimeography Pro < 0.7 run some code
that threw errors around. The biggest change was that most variables are no longer
manipulated in the constructor (as it is bad practice.)

Vimeography Pro < 0.7 used several methods that were making changes to these variables,
so I've introduced some methods that now call an ugly manual reload on a few Pro class constructors
to ensure that the necessary Vimeography Pro parameters are set _after_ the variable is
actually set up. Otherwise, Pro tries to set the variable data before it even exists, so the Pro
data is overwritten when the _actual_ object is set up.

Also, the Vimeo PHP library is now namespaced, so I threw in a class_alias shim to try and capture
any calls to the Vimeo PHP library that are un-namespaced.

The Vimeography_Deprecated class is a good idea for future changes that will allow for backwards-compatibility.

We'll also perform the common template-loading methods inside of the renderer in a new public method
called `load_theme` - this way, we don't have to duplicate the same actions inside of Vimeography Pro's
render function.

Lastly, introducing the deprecated `$_theme` instance variable ensures that Vimeography Pro < 0.7 does not fail when the
rendered tries to render `$_theme` instead of `$_template`