Refactored the gallery list page
Instead of using the home-rolled table, we now extend from WP_List_Table class to use native WordPress helper functions to display our data. This allows us to take advantage of sortable columns, easier pagination, native styles etc.

We also got rid of the gallery quickview, because I'm pretty sure the only useful data in that dropdown was the theme that is applied to the gallery. That information now has its own column in the table.

Also included in this commit is new duplicate gallery functionality. If the user elects to duplicate a gallery, they will now be presented with a jQuery dialog box that allows them to define a new name and source for their gallery.

The default values are the same as what they were for the gallery that they chose to duplicate, but can be edited to whatever the user would like.

There is also a checkbox that asks if the user would like to also copy over any theme appearance customizations that the user may have made. If this is selected, the original gallery's CSS file will be read, parsed, and saved as a CSS stylesheet for the new gallery, using the new gallery's ID.