Add compatibility with WP Multisite
We shouldn't be saving the table name to a constant.

If you're on multisite, a call to the `switch_to_blog()` function
would continue to load the first site's database.

This change will ensure we use the right database prefix on every
Vimeography database call, regardless of which site we are on.