Adding some mock code for 2.0 idea
The thought here is that Vimeography 2.0 themes will register an
entrypoint (labeled `app_js` in this commit) that Vimeography 2.0 can detect.

If this entrypoint exists, we know the theme supports the new 2.0 loading mechanism.

At this point, we'd need to also be sure that the installed version of Vimeography PRO
also supports loading the theme as a React app instead of rendering it on the server side.

We still need to determine what to do if a theme is 2.0-ready, but Vimeography PRO is not.

Same of the inverse situation, where Vimeography PRO is 2.0-ready but the theme is not.
IE. should the theme fall back to being rendered on the server? Or should it die entirely?

This is a proof of concept, and none of this code is meant to make it past development.