Refactor Onboarding Process
Instead of presenting the users with three options to choose their own adventure,
we'll guide them through creating their first gallery as soon as the plugin
is activated.

This allows the user to see their videos inside the plugin as quickly as possible.

We can then offer a way in which the user can quickly preview their videos in our
other gallery themes.