Change prefix of gallery wrapper class
Without doing this, all CSS customizations that were made before the
new theme markup was introduced might still be applied to the new HTML markup.

For example, if an old `<div>` had a class of `.vimeography-aloha .vimeography-thumbnail`,
its CSS customization would still apply to the new HTML  markup, even if the customization
doesn't fit the requirements of the new HTML markup.