Add a filter to output a referrer policy
Some Vimeo videos with domain restrictions will not display unless
this referrer policy is enabled on the WordPress site, but it's
probably a good idea to have the user opt in to using it rather
than enabling it by default for all sites.

If a domain restricted video is not displaying on your site, you
may need to enable this option.