Appearance editing 2.0
This commit adds endpoints for persisting custom CSS via the REST API.
The CSS is stored using WP functions instead of writing to disk
like we used to do.

When the CSS stylesheet's ID is detected in the admin, we parse
the style and convert all of the values into rules that are pushed
into state. This allows us to populate the controls with the
existing user customizations rather than the default values
provided in the theme's settings.php file